[canvas] Create Nav Link service; remove legacy service (#107346)
