+++ x-pack/plugins/canvas/public/services/index.ts
@@ -9,6 +9,7 @@ export * from './legacy';
 
 import { PluginServices } from '../../../../../src/plugins/presentation_util/public';
 import { CanvasWorkpadService } from './workpad';
+import { CanvasNavLinkService } from './nav_link';
 import { CanvasNotifyService } from './notify';
 import { CanvasPlatformService } from './platform';
 
@@ -16,10 +17,12 @@ export interface CanvasPluginServices {
   workpad: CanvasWorkpadService;
   notify: CanvasNotifyService;
   platform: CanvasPlatformService;
+  navLink: CanvasNavLinkService;
 }
 
 export const pluginServices = new PluginServices<CanvasPluginServices>();
 
 export const useWorkpadService = () => (() => pluginServices.getHooks().workpad.useService())();
+export const useNavLinkService = () => (() => pluginServices.getHooks().navLink.useService())();
 export const useNotifyService = () => (() => pluginServices.getHooks().notify.useService())();
 export const usePlatformService = () => (() => pluginServices.getHooks().platform.useService())();
