+++ x-pack/plugins/canvas/public/services/legacy/context.tsx
@@ -22,7 +22,6 @@ export interface WithServicesProps {
 const defaultContextValue = {
   embeddables: {},
   expressions: {},
-  navLink: {},
   search: {},
 };
 
@@ -31,7 +30,6 @@ export const ServicesContext = createContext<CanvasServices>(defaultContextValue
 export const useServices = () => useContext(ServicesContext);
 export const useEmbeddablesService = () => useServices().embeddables;
 export const useExpressionsService = () => useServices().expressions;
-export const useNavLinkService = () => useServices().navLink;
 export const useLabsService = () => useServices().labs;
 export const useReportingService = () => useServices().reporting;
 
@@ -49,7 +47,6 @@ export const LegacyServicesProvider: FC<{
   const value = {
     embeddables: specifiedProviders.embeddables.getService(),
     expressions: specifiedProviders.expressions.getService(),
-    navLink: specifiedProviders.navLink.getService(),
     search: specifiedProviders.search.getService(),
     reporting: specifiedProviders.reporting.getService(),
     labs: specifiedProviders.labs.getService(),
