+++ x-pack/plugins/canvas/public/services/legacy/index.ts
@@ -8,7 +8,6 @@
 import { BehaviorSubject } from 'rxjs';
 import { CoreSetup, CoreStart, AppUpdater } from '../../../../../../src/core/public';
 import { CanvasSetupDeps, CanvasStartDeps } from '../../plugin';
-import { navLinkServiceFactory } from './nav_link';
 import { embeddablesServiceFactory } from './embeddables';
 import { expressionsServiceFactory } from './expressions';
 import { searchServiceFactory } from './search';
@@ -16,7 +15,6 @@ import { labsServiceFactory } from './labs';
 import { reportingServiceFactory } from './reporting';
 
 export { SearchService } from './search';
-export { NavLinkService } from './nav_link';
 export { EmbeddablesService } from './embeddables';
 export { ExpressionsService } from '../../../../../../src/plugins/expressions/common';
 export * from './context';
@@ -75,7 +73,6 @@ export type ServiceFromProvider<P> = P extends CanvasServiceProvider<infer T> ? 
 export const services = {
   embeddables: new CanvasServiceProvider(embeddablesServiceFactory),
   expressions: new CanvasServiceProvider(expressionsServiceFactory),
-  navLink: new CanvasServiceProvider(navLinkServiceFactory),
   search: new CanvasServiceProvider(searchServiceFactory),
   reporting: new CanvasServiceProvider(reportingServiceFactory),
   labs: new CanvasServiceProvider(labsServiceFactory),
@@ -86,7 +83,6 @@ export type CanvasServiceProviders = typeof services;
 export interface CanvasServices {
   embeddables: ServiceFromProvider<typeof services.embeddables>;
   expressions: ServiceFromProvider<typeof services.expressions>;
-  navLink: ServiceFromProvider<typeof services.navLink>;
   search: ServiceFromProvider<typeof services.search>;
   reporting: ServiceFromProvider<typeof services.reporting>;
   labs: ServiceFromProvider<typeof services.labs>;
@@ -112,7 +108,6 @@ export const stopServices = () => {
 
 export const {
   embeddables: embeddableService,
-  navLink: navLinkService,
   expressions: expressionsService,
   search: searchService,
   reporting: reportingService,
