+++ x-pack/plugins/canvas/public/services/legacy/stubs/index.ts
@@ -9,7 +9,6 @@ import { CanvasServices, services } from '../';
 import { embeddablesService } from './embeddables';
 import { expressionsService } from './expressions';
 import { reportingService } from './reporting';
-import { navLinkService } from './nav_link';
 import { labsService } from './labs';
 import { searchService } from './search';
 
@@ -17,7 +16,6 @@ export const stubs: CanvasServices = {
   embeddables: embeddablesService,
   expressions: expressionsService,
   reporting: reportingService,
-  navLink: navLinkService,
   search: searchService,
   labs: labsService,
 };
